// backend/src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'super_admin',
  BRAND_OWNER = 'brand_owner',
  BRANCH_MANAGER = 'branch_manager',
  BRANCH_WAITER = 'branch_waiter',
  BRANCH_CHEF = 'branch_chef',
  RESTAURANT_OWNER = 'restaurant_owner',
  MANAGER = 'manager',
  WAITER = 'waiter',
  CHEF = 'chef',
  CUSTOMER = 'customer',
}