// backend/src/modules/audit/enums/audit-action.enum.ts
export enum AuditAction {
  SETTING_UPDATED = 'SETTING_UPDATED',
  CASH_REGISTER_CREATED = 'CASH_REGISTER_CREATED',
  CASH_REGISTER_DELETED = 'CASH_REGISTER_DELETED',
  CASH_DEFAULT_REGISTER_ENSURED = 'CASH_DEFAULT_REGISTER_ENSURED',
  CASH_SESSION_OPENED = 'CASH_SESSION_OPENED',
  CASH_SESSION_CLOSED = 'CASH_SESSION_CLOSED',
  CASH_MOVEMENT_ADDED = 'CASH_MOVEMENT_ADDED',
  INVENTORY_INGREDIENT_CREATED = 'INVENTORY_INGREDIENT_CREATED',
  INVENTORY_INGREDIENT_UPDATED = 'INVENTORY_INGREDIENT_UPDATED',
  INVENTORY_INGREDIENT_DELETED = 'INVENTORY_INGREDIENT_DELETED',
  INVENTORY_STOCK_MOVEMENT_ADDED = 'INVENTORY_STOCK_MOVEMENT_ADDED',
  INVENTORY_RECIPE_CREATED = 'INVENTORY_RECIPE_CREATED',
  INVENTORY_RECIPE_DELETED = 'INVENTORY_RECIPE_DELETED',
  INVENTORY_STOCK_BULK_UPDATED = 'INVENTORY_STOCK_BULK_UPDATED',
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_STATUS_UPDATED = 'ORDER_STATUS_UPDATED',
  ORDER_STATUS_BATCH_UPDATED = 'ORDER_STATUS_BATCH_UPDATED',
  ORDER_MOVED_TO_TABLE = 'ORDER_MOVED_TO_TABLE',
  ORDER_ITEMS_UPDATED = 'ORDER_ITEMS_UPDATED',
  PAYMENT_CREATED = 'PAYMENT_CREATED',
  PAYMENT_SPLIT_CREATED = 'PAYMENT_SPLIT_CREATED',
  PAYMENT_REVERTED = 'PAYMENT_REVERTED',
  CUSTOMER_CREATED = 'CUSTOMER_CREATED',
  CUSTOMER_UPDATED = 'CUSTOMER_UPDATED',
  CUSTOMER_DELETED = 'CUSTOMER_DELETED',
  RESERVATION_CREATED = 'RESERVATION_CREATED',
  RESERVATION_UPDATED = 'RESERVATION_UPDATED',
  RESERVATION_STATUS_UPDATED = 'RESERVATION_STATUS_UPDATED',
  RESTAURANT_CREATED = 'RESTAURANT_CREATED',
  RESTAURANT_UPDATED = 'RESTAURANT_UPDATED',
  RESTAURANT_DELETED = 'RESTAURANT_DELETED',
  BRANCH_CATEGORY_HIDDEN = 'BRANCH_CATEGORY_HIDDEN',
  BRANCH_CATEGORY_SHOWN = 'BRANCH_CATEGORY_SHOWN',
  BRANCH_CATEGORIES_HIDDEN_BULK = 'BRANCH_CATEGORIES_HIDDEN_BULK',
  BRANCH_CATEGORIES_SHOWN_BULK = 'BRANCH_CATEGORIES_SHOWN_BULK',
  BRANCH_ITEM_OVERRIDE_UPSERTED = 'BRANCH_ITEM_OVERRIDE_UPSERTED',
  BRANCH_ITEM_OVERRIDE_REMOVED = 'BRANCH_ITEM_OVERRIDE_REMOVED',
  BRANCH_ITEM_OVERRIDE_BULK_APPLIED = 'BRANCH_ITEM_OVERRIDE_BULK_APPLIED',
  MENU_CATEGORY_CREATED = 'MENU_CATEGORY_CREATED',
  MENU_CATEGORY_UPDATED = 'MENU_CATEGORY_UPDATED',
  MENU_CATEGORY_DELETED = 'MENU_CATEGORY_DELETED',
  MENU_ITEM_CREATED = 'MENU_ITEM_CREATED',
  MENU_ITEM_UPDATED = 'MENU_ITEM_UPDATED',
  MENU_ITEM_DELETED = 'MENU_ITEM_DELETED',
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_STATUS_CHANGED = 'USER_STATUS_CHANGED',
  USER_DELETED = 'USER_DELETED',
  USER_RESTORED = 'USER_RESTORED',
  USER_ASSIGNED_RESTAURANT = 'USER_ASSIGNED_RESTAURANT',
}